requirements.txt
# Core Django
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1



# Security & Authentication
PyJWT==2.10.1

# CORS
django-cors-headers==4.7.0

# Static Files
whitenoise==6.11.0

# Admin UI
django-jazzmin

# Media/Image Processing
pillow==11.3.0

# Server
gunicorn==23.0.0

# Configuration
python-decouple==3.8

# Telegram Bot (se necessário)
python-telegram-bot==22.3
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0

# Utilities
asgiref==3.9.1
certifi==2025.8.3
idna==3.10
packaging==25.0
sniffio==1.3.1
sqlparse==0.5.3
tzdata==2025.2
anyio==4.10.0

# Production Dependencies
django-filter==24.3